Move PipelineFile location due to deps
