+++ pipeline/objects/__init__.py
@@ -3,8 +3,7 @@ from pipeline.objects.function import Function
 from pipeline.objects.graph import Graph
 from pipeline.objects.model import Model
 from pipeline.objects.pipeline import Pipeline
-from pipeline.objects.variable import Variable
-from pipeline.objects.file import PipelineFile
+from pipeline.objects.variable import Variable, PipelineFile
 
 __all__ = [
     "Pipeline",
