+++ pipeline/objects/variable.py
@@ -43,3 +43,24 @@ class Variable:
             name=schema.name,
             local_id=schema.local_id,
         )
+
+
+class PipelineFile(Variable):
+
+    path: str
+
+    def __init__(
+        self,
+        *,
+        path: str = None,
+        name: str = None,
+        remote_id: str = None,
+    ) -> None:
+        super().__init__(
+            type_class=self,
+            is_input=False,
+            is_output=False,
+            name=name,
+            remote_id=remote_id,
+        )
+        self.path = path
